회원 인증 서비스 인터페이스 생성
- 회원 인증을 담당하는 인터페이스를 생성합니다.
- 기존 서비스 클래스의 이름을 AuthenticationService -> UserLoginService로 변경합니다.
